+++ changelog.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - charmed
     - stunned
     - unconscious
+- ability to toggle pawn visibility
 
 ### Fixed